+++ package.json
@@ -26,9 +26,7 @@
     "lint": "prettier --check '{src,test}/**/*.js' && eslint '{src,test}/**/*.js'",
     "pretest": "npm run compile && npm run lint",
     "readme": "doctoc --maxlevel 3 README.md",
-    "test": "npm run test:node && npm run test:browser",
-    "test:node": "NODE_ENV=testing nyc mocha",
-    "test:browser": "NODE_ENV=testing karma start"
+    "test": "jest --coverage"
   },
   "dependencies": {
     "axios": "^0.19.2",
@@ -54,7 +52,6 @@
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "babel-plugin-istanbul": "^6.0.0",
-    "chai": "^4.2.0",
     "compression-webpack-plugin": "^3.1.0",
     "coveralls": "^3.0.11",
     "css-loader": "^3.4.2",
@@ -71,24 +68,13 @@
     "file-loader": "^6.0.0",
     "http-server": "^0.12.1",
     "husky": "^4.2.3",
-    "jsdom": "^15.1.1",
-    "karma": "^1.6.0",
-    "karma-jsdom-launcher": "^8.0.0",
-    "karma-mocha": "^1.3.0",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-sinon-chai": "^1.2.4",
-    "karma-spec-reporter": "0.0.26",
-    "karma-webpack": "^2.0.2",
+    "jest": "^25.2.4",
     "mini-css-extract-plugin": "^0.9.0",
-    "mocha": "^6.2.1",
     "npm-check-updates": "^4.1.0",
-    "nyc": "^14.1.1",
     "prettier": "^2.0.2",
     "react-hot-loader": "^4.12.20",
     "react-test-renderer": "^16.13.1",
     "referencer": "^0.2.5",
-    "sinon": "^2.1.0",
-    "sinon-chai": "^2.9.0",
     "style-loader": "^1.1.3",
     "webpack": "^4.42.1",
     "webpack-cli": "^3.3.11",
