--- test/karma.opts
@@ -1,2 +1,0 @@
-const testsContext = require.context('.', true, /\.js$/)
-testsContext.keys().forEach(testsContext)
