--- test/mocha.opts
@@ -1,3 +1,0 @@
---require @babel/register
---timeout 10000
---recursive
